Clean up, change paths from relative to absolute. Move types to utils, delete not needed vendor files
