+++ src/pages/404/404.styles.ts
@@ -2,7 +2,7 @@ import styled from 'styled-components/macro';
 import { spacing } from '@material-ui/system';
 import { Button as MuiButton } from '@material-ui/core';
 
-import { MuiButtonSpacingType } from '../../types/styles';
+import { MuiButtonSpacingType } from '@utils/types/styles';
 
 export const Button = styled(MuiButton)<MuiButtonSpacingType>(spacing);
