+++ src/utils/types/styles.ts
